real->double

+++ src/config.hpp
@@ -20,33 +20,33 @@ typedef double REAL;
 typedef unsigned int uint;
 
 // Physical constants
-//const REAL LAMBDAQCD2 = 0.21416*0.21416;   // GeV^2
-const REAL LAMBDAQCD2 = 0.241*0.241;    // 0902.1112
-const REAL LAMBDAQCD = 0.241;
-//const REAL LAMBDAQCD2 = 0.2*0.2;    // 0704.0612, AN06 model
+//const double LAMBDAQCD2 = 0.21416*0.21416;   // GeV^2
+const double LAMBDAQCD2 = 0.241*0.241;    // 0902.1112
+const double LAMBDAQCD = 0.241;
+//const double LAMBDAQCD2 = 0.2*0.2;    // 0704.0612, AN06 model
 const int Nf=3;
 const int Nc=3;
-const REAL Cf = (Nc*Nc-1.0)/(2.0*Nc);
-const REAL ALPHA_e = 1.0/137.035999679; 
-const REAL e = sqrt(4.0*M_PI*ALPHA_e);
+const double Cf = (Nc*Nc-1.0)/(2.0*Nc);
+const double ALPHA_e = 1.0/137.035999679; 
+const double e = sqrt(4.0*M_PI*ALPHA_e);
 
 // Reqularization of the running coupling
 // Berger&Stasto 1010.0671: 0.3
 // Fit to HERA: 0.7 0902.1112
-//const REAL MAXALPHA = 0.7;  
-const REAL MAXALPHA = 0.5;  // 0704.0612
-const REAL ALPHABAR_s = 0.05;       // \alphabar_s if RC=constant
-const REAL ALPHAS = 0.2;    // compare with ALPHABAR_s.... ok, doesn't make sense
+//const double MAXALPHA = 0.7;  
+const double MAXALPHA = 0.5;  // 0704.0612
+const double ALPHABAR_s = 0.05;       // \alphabar_s if RC=constant
+const double ALPHAS = 0.2;    // compare with ALPHABAR_s.... ok, doesn't make sense
 
 // Other constants
 
-const REAL eps=0.000001;
+const double eps=0.000001;
 
 // Inline functions
 
 #define LINEINFO __FILE__ << ":" << __LINE__
 
-inline const REAL SQR(const REAL x) { return x*x; }
+inline const double SQR(const double x) { return x*x; }
 
 enum Parton
 {
