+++ src/pdf/mrst.hpp
@@ -17,7 +17,7 @@ class MRST : public PDF
 {
     public:
         ~MRST();
-        REAL xq(REAL x, REAL q, Parton p);    // return x*q(x,q)
+        double xq(double x, double q, Parton p);    // return x*q(x,q)
         void Initialize(int param=-1);
         std::string GetString();
     private:
